requirements.txt
langchain
openai
streamlit
tiktoken
pinecone-client
python-decouple
transformers
sentence-transformers
qdrant-client
pytest